resume-optimizer-api-deployable/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-multipart==0.0.9
python-docx==1.1.2
pdfminer.six==20231228
reportlab==4.2.0
requests==2.32.3
jinja2==3.1.4
